Make source analyses handle epoched inputs
Former-commit-id: 899e5a83a0fde5526af5287fa44cc762b811f12b
